Rename Yarp.Common.Tests -> Yarp.Tests.Common
