+++ test/ReverseProxy.Tests/Common/TaskExtensions.cs
@@ -10,7 +10,7 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Yarp.Common.Tests
+namespace Yarp.Tests.Common
 {
     /// <summary>
     /// Extensions for the <see cref="Task"/> class.
