+++ test/ReverseProxy.Tests/Transforms/RequestHeaderRemoveTransformTests.cs
@@ -7,7 +7,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Xunit;
-using Yarp.Common.Tests;
+using Yarp.Tests.Common;
 using Yarp.ReverseProxy.Utilities.Tests;
 
 namespace Yarp.ReverseProxy.Transforms.Tests
