+++ test/ServiceFabric.Tests/ServiceDiscovery/Util/CachedServiceFabricCallerTests.cs
@@ -11,7 +11,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using Xunit;
 using Xunit.Abstractions;
-using Yarp.Common.Tests;
+using Yarp.Tests.Common;
 using Yarp.ReverseProxy.Utilities;
 
 namespace Yarp.ReverseProxy.ServiceFabric.Tests
